fix type of rating

+++ CyberSwift/Models/ResponseAPI/ResponseAPIContentCommunity.swift
@@ -44,7 +44,7 @@ public struct ResponseAPIContentGetLeaders: Decodable {
 
 public struct ResponseAPIContentGetLeader: Decodable, Equatable {
     public let url: String
-    public let rating: UInt
+    public let rating: Double
     public let isActive: Bool
     public let userId: String
     public let position: UInt64
